Move Metadata base classes into utils
