+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/Event.java
@@ -3,7 +3,7 @@ package com.dat3m.dartagnan.program.event;
 import com.dat3m.dartagnan.program.Function;
 import com.dat3m.dartagnan.program.Program;
 import com.dat3m.dartagnan.program.Thread;
-import com.dat3m.dartagnan.program.event.metadata.Metadata;
+import com.dat3m.dartagnan.utils.metadata.Metadata;
 import com.dat3m.dartagnan.verification.Context;
 
 import java.util.Collection;
