+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/metadata/CompilationId.java
@@ -1,4 +1,6 @@
 package com.dat3m.dartagnan.program.event.metadata;
 
+import com.dat3m.dartagnan.utils.metadata.Metadata;
+
 // Used as a snapshot of the global ID right before compilation.
 public record CompilationId(int value) implements Metadata { }
