+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/metadata/CustomPrinting.java
@@ -2,6 +2,7 @@ package com.dat3m.dartagnan.program.event.metadata;
 
 
 import com.dat3m.dartagnan.program.event.Event;
+import com.dat3m.dartagnan.utils.metadata.Metadata;
 
 import java.util.Optional;
