--- dartagnan/src/main/java/com/dat3m/dartagnan/program/event/metadata/Metadata.java
@@ -1,4 +1,0 @@
-package com.dat3m.dartagnan.program.event.metadata;
-
-public interface Metadata {
-}
