+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/metadata/OriginalId.java
@@ -1,4 +1,6 @@
 package com.dat3m.dartagnan.program.event.metadata;
 
+import com.dat3m.dartagnan.utils.metadata.Metadata;
+
 // Used as a snapshot of the global ID after the program has been constructed (either programmatically or via a parser).
 public record OriginalId(int value) implements Metadata { }
