+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/metadata/UnrollingBound.java
@@ -1,3 +1,5 @@
 package com.dat3m.dartagnan.program.event.metadata;
 
+import com.dat3m.dartagnan.utils.metadata.Metadata;
+
 public record UnrollingBound(int value) implements Metadata { }
