+++ dartagnan/src/main/java/com/dat3m/dartagnan/utils/metadata/Metadata.java
@@ -1,0 +1,4 @@
+package com.dat3m.dartagnan.utils.metadata;
+
+public interface Metadata {
+}
